requirements.txt
pymupdf==1.22.3
aioredis==2.0.1
loguru==0.7.0
peewee==3.16.2
tiktoken==0.4.0
aiofiles==23.1.0
httpx==0.24.1
pymysql2==1.3.3
python-dotenv==1.0.0
flask==2.3.2
numpy==1.24.3
beautifulsoup4==4.12.2
pillow==10.0.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
crcmod==1.7
jmespath==0.10.0
oss2==2.18.0

